cleaned up spherical harmonics by merging two files
